backend/requirements.txt
annotated-doc==0.0.4
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.37.0
scipy>=1.10.0
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
coverage==7.13.4
dnspython==2.8.0
docopt==0.6.2
email-validator==2.3.0
fastapi==0.128.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
minio==7.2.20
packaging==26.0
passlib==1.7.4
pillow==11.1.0
pipreqs==0.4.13
pluggy==1.6.0
psycopg2-binary==2.9.11
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-dotenv==1.2.1
python-multipart==0.0.20
pypdf==5.1.0
PyYAML==6.0.3
requests==2.32.5
SQLAlchemy==2.0.47
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1; sys_platform != 'win32'
watchfiles==1.1.1
websockets==16.0
yarg==0.1.10
